Revert "revert"
This reverts commit bc61b7ea64ba4ff7e796df49d91757798e75abdf.

+++ src/Point.java
@@ -9,7 +9,7 @@ import java.awt.Toolkit;
 public class Point
 {
     private double x, y, z, depthScale;
-    private double twoDX, twoDY;
+    private int twoDX, twoDY;
     private static final double WIDTH = Toolkit.getDefaultToolkit().getScreenSize().getWidth() / 2;
 
     public Point(double x, double y, double z) {
@@ -29,15 +29,15 @@ public class Point
         y = newY;
         z = newZ;
         depthScale = WIDTH / (20 + z);
-        twoDX = (double)(depthScale * x);
-        twoDY = (double)(depthScale * y);
+        twoDX = (int)(depthScale * x);
+        twoDY = (int)(depthScale * y);
     }
 
-    public double get2Dx() {
+    public int get2Dx() {
         return twoDX;
     }
 
-    public double get2Dy() {
+    public int get2Dy() {
         return twoDY;
     }
